added swagger OpenApi generator function

+++ Event_system/Migrations/EventDbContextModelSnapshot.cs
@@ -61,8 +61,8 @@ namespace Event_system.Migrations
                     b.Property<DateTime>("EndDate")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<int>("Price")
-                        .HasColumnType("integer");
+                    b.Property<decimal>("Price")
+                        .HasColumnType("numeric");
 
                     b.Property<DateTime>("StartDate")
                         .HasColumnType("timestamp with time zone");
